+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.4.1
+
 * Added Python 3.11 support. 
 
 # 0.4.0