+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.4.1
+
+* Added Python 3.11 support. 
+
 # 0.4.0
 
 ## Major changes